requirements.txt
Django==4.1
django-tenants==3.6.1
django-cors-headers
psycopg2
djangorestframework
markdown
django-filter
django-auditlog